#532 - move integer BigIntegerField

+++ netbox_custom_objects/field_types.py
@@ -366,7 +366,10 @@ class IntegerFieldType(FieldType):
         # TODO: handle all args for IntegerField
         field_kwargs = self._safe_kwargs(**kwargs)
         field_kwargs.update({"default": field.default, "unique": field.unique})
-        return models.IntegerField(null=True, blank=True, **field_kwargs)
+        # Use a 64-bit column (bigint) so values are not capped at the 32-bit
+        # signed range. Matches NetBox core's convention for semantic integers
+        # (e.g. ASNField, L2VPN identifier). See issue #532.
+        return models.BigIntegerField(null=True, blank=True, **field_kwargs)
 
     def get_filterform_field(self, field, **kwargs):
         return forms.IntegerField(
